[MIPR-1367] Initial Refactor to create a separate IndexController
