+++ app/uk/gov/hmrc/incometaxpenaltiesfrontend/controllers/LanguageSwitchController.scala
@@ -31,5 +31,5 @@ class LanguageSwitchController @Inject()(
     "cymraeg" -> Lang("cy")
   )
 
-  override def fallbackURL: String = routes.ServiceController.homePage.url
+  override def fallbackURL: String = routes.IndexController.homePage.url
 }
