+++ app/uk/gov/hmrc/incometaxpenaltiesfrontend/views/PenaltyCalculation.scala.html
@@ -72,7 +72,7 @@
   ))
 
     @link(
-        link = uk.gov.hmrc.incometaxpenaltiesfrontend.controllers.routes.ServiceController.homePage.url,
+        link = uk.gov.hmrc.incometaxpenaltiesfrontend.controllers.routes.IndexController.homePage.url,
         messageKey = messages("calculation.return.link"),
         id = Some("returnToIndex")
     )
