+++ conf/app.routes
@@ -5,7 +5,7 @@
 GET        /assets/*file                        controllers.Assets.versioned(path = "/public", file: Asset)
 
 
-GET        /                                    uk.gov.hmrc.incometaxpenaltiesfrontend.controllers.ServiceController.homePage
+GET        /                                    uk.gov.hmrc.incometaxpenaltiesfrontend.controllers.IndexController.homePage
 
 GET        /sign-out                            uk.gov.hmrc.incometaxpenaltiesfrontend.controllers.ServiceController.signOut()
 GET        /keep-alive                          uk.gov.hmrc.incometaxpenaltiesfrontend.controllers.ServiceController.keepAlive()
