+++ it/test/uk/gov/hmrc/incometaxpenaltiesfrontend/controllers/ServiceControllerISpec.scala
@@ -27,69 +27,6 @@ import java.net.URLEncoder
 
 class ServiceControllerISpec extends ComponentSpecHelper with ViewSpecHelper with AuthStub {
 
-  "GET /" should {
-    "return an OK with a view" when {
-      "the user is an authorised individual" in {
-        stubAuth(OK, successfulIndividualAuthResponse)
-        val result = get("/")
-
-        result.status shouldBe OK
-      }
-
-      "the user is an authorised agent" in {
-        stubAuth(OK, successfulAgentAuthResponse)
-        val result = get("/", isAgent = true)
-
-        result.status shouldBe OK
-      }
-    }
-    "have the correct page has correct elements" when {
-      "the user is an authorised individual" in {
-        stubAuth(OK, successfulIndividualAuthResponse)
-        val result = get("/")
-
-        val document = Jsoup.parse(result.body)
-
-        document.getServiceName.text() shouldBe "Manage your Self Assessment"
-        document.title() shouldBe "Self Assessment penalties and appeals - Manage your Self Assessment - GOV.UK"
-        document.getH1Elements.text() shouldBe "Self Assessment penalties and appeals"
-        document.getH2Elements.get(0).text() shouldBe "Overview"
-        document.getParagraphs.get(0).text() shouldBe "Your account has:"
-        document.getH2Elements.get(1).text() shouldBe "Penalty and appeal details"
-        document.getH3Elements.get(0).text() shouldBe "Late submission penalties"
-        document.getH3Elements.get(1).text() shouldBe "Late payment penalties"
-        document.getElementById("financialWarningText").text() shouldBe "! Warning You have reached the financial penalty threshold."
-        document.getSubmitButton.text() shouldBe "Check amounts and pay"
-        document.getLink("actionsToRemoveLink").text() shouldBe "Actions to take to get your points removed (opens in new tab)"
-        document.getLink("actionsToRemoveLink").attr("href") shouldBe "/penalties/income-tax/compliance-timeline"
-        document.getLink("appealPenalty1Link").attr("href") shouldBe "/penalties/income-tax/appeal-penalty?penaltyId=1"
-        document.getLink("appealPenaltyPoint1Link").attr("href") shouldBe "/penalties/income-tax/appeal-penalty?penaltyId=1"
-      }
-
-      "the user is an authorised agent" in {
-        stubAuth(OK, successfulAgentAuthResponse)
-        val result = get("/", isAgent = true)
-
-        val document = Jsoup.parse(result.body)
-
-        document.getServiceName.text() shouldBe "Manage your Self Assessment"
-        document.title() shouldBe "Self Assessment penalties and appeals - Manage your Self Assessment - GOV.UK"
-        document.getH1Elements.text() shouldBe "Self Assessment penalties and appeals"
-        document.getH2Elements.get(0).text() shouldBe "Overview"
-        document.getParagraphs.get(0).text() shouldBe "Your client's account has:"
-        document.getH2Elements.get(1).text() shouldBe "Penalty and appeal details"
-        document.getH3Elements.get(0).text() shouldBe "Late submission penalties"
-        document.getH3Elements.get(1).text() shouldBe "Late payment penalties"
-        document.getElementById("financialWarningText").text() shouldBe "! Warning Your client has been given a £200 penalty for reaching the penalty threshold."
-        document.getSubmitButton.text() shouldBe "Check amounts and pay"
-        document.getLink("actionsToRemoveLink").text() shouldBe "Actions your client must take to get their points removed (opens in new tab)"
-        document.getLink("actionsToRemoveLink").attr("href") shouldBe "/penalties/income-tax/compliance-timeline"
-        document.getLink("appealPenalty1Link").attr("href") shouldBe "/penalties/income-tax/appeal-penalty?penaltyId=1"
-        document.getLink("appealPenaltyPoint1Link").attr("href") shouldBe "/penalties/income-tax/appeal-penalty?penaltyId=1"
-      }
-    }
-  }
-
   "GET /penalties/income-tax/sign-out" should {
     "redirect to sign-out route with the continue URL set to the feedback survey" in {
       val appConfig = app.injector.instanceOf[AppConfig]
